requirements.txt
ipykernel
numpy
pytorch
requests
